Move PsychicInventory and SleepManager storage into their own class
